add shared/memberships-edit.js
